% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{package}
\name{BRSPackage}
\alias{BRSPackage}
\alias{BRSPackage-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
BRSPackage
A data package for BRSPackage.

Use \code{data(package='BRSPackage')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{series_range}
}
