% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{cee19_01}
\alias{cee19_01}
\title{Descriptive data title}
\format{
a \code{data.frame} containing the following fields:
\describe{
\item{DeployID}{}
\item{start_date}{}
\item{end_date}{}
\item{num_fixes}{}
\item{date_range}{}
\item{average_locs}{}
\item{series_start_date}{}
\item{series_end_date}{}
\item{series_date_range}{}
\item{aux}{}
\item{gonio}{}
\item{locations}{}
\item{num_before_cee}{}
\item{num_after_cee}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{BRSPackage}
\link{cee19_02}
\link{cee19_03}
\link{cee19_04}
}
