% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{gonio_out}
\alias{gonio_out}
\title{Collapsed gonio data}
\format{
a \code{data.frame} containing the following fields:
\describe{
\item{DeployID}{The ID identifying an individual whale.}
\item{date_time}{The date and time of an observation.}
\item{lat}{The latitude where an individual whale was observed.}
\item{lon}{The longitude where an individual whale was observed.}
\item{platform}{The vessel the observation was made from.}
\item{distance}{A measure of the error of the observation.}
\item{Quality}{The category based on the magnitude of error of an observation.}
\item{Source}{Whether an observation is from locations, auxiliary, or goniometer data.}
}
}
\source{
The data comes from the goniometer data from the 2019 BRS Study.
}
\description{
Detailed description of the data
}
\seealso{

}
