% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{locations}
\alias{locations}
\title{Descriptive data title}
\format{
a \code{spec_tbl_df} containing the following fields:
\describe{
\item{X1}{}
\item{DeployID}{}
\item{Ptt}{}
\item{Instr}{}
\item{Date}{}
\item{Type}{}
\item{Quality}{}
\item{Latitude}{}
\item{Longitude}{}
\item{Error.radius}{}
\item{Error.Semi.major.axis}{}
\item{Error.Semi.minor.axis}{}
\item{Error.Ellipse.orientation}{}
\item{Offset}{}
\item{Offset.orientation}{}
\item{GPE.MSD}{}
\item{GPE.U}{}
\item{Count}{}
\item{Comment}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{BRSPackage}
}
