% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{series}
\alias{series}
\title{Dive data of whales}
\format{
a \code{data.frame} containing the following fields:
\describe{
\item{DeployID}{The ID identifying an individual whale.}
\item{Ptt}{The number identifying an individual whale.}
\item{DepthSensor}{The depth of each observation.}
\item{Source}{}
\item{Instr}{Instrument used to identify an individual whale.}
\item{Day}{The date of the first observation from an individual whale.}
\item{Time}{The time of the first observation from an individual whale.}
\item{LocationQuality}{N/A}
\item{Latitude}{N/A}
\item{Longitude}{N/A}
\item{Depth}{The depth of each individual observation.}
\item{DRange}{The depth range of each individual observation.}
\item{Temperature}{N/A}
\item{TRange}{N/A}
\item{Activity}{N/A}
\item{ARange}{N/A}
}
}
\source{
The data comes from the 2019 BRS Study.
}
\description{
Detailed description of the data
}
\seealso{

}
