% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{series_range}
\alias{series_range}
\title{Descriptive data title}
\format{
a \code{spec_tbl_df} containing the following fields:
\describe{
\item{X1}{}
\item{DeployID}{}
\item{Ptt}{}
\item{DepthSensor}{}
\item{Source}{}
\item{Instr}{}
\item{Count}{}
\item{Start}{}
\item{End}{}
\item{LocationQuality}{}
\item{Latitude}{}
\item{Longitude}{}
\item{MinDepth}{}
\item{MinDepthAccuracy}{}
\item{MaxDepth}{}
\item{MaxDepthAccuracy}{}
\item{MinTemp}{}
\item{MinTempAccuracy}{}
\item{MaxTemp}{}
\item{MaxTempAccuracy}{}
\item{MobMean}{}
\item{MobSD}{}
\item{ActivitySum}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{BRSPackage}
}
