% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{series_xts}
\alias{series_xts}
\title{Descriptive data title}
\format{
a \code{xts} containing the following fields:
\describe{
\item{depth}{}
\item{lower}{}
\item{upper}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{BRSPackage}
\link{series}
}
