% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{aux}
\alias{aux}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{DeployID}{}
\item{date_time}{}
\item{lat}{}
\item{lon}{}
\item{platform}{}
\item{distance}{}
\item{Quality}{}
\item{Source}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{BRSPackage}
\link{gonio}
\link{processed_gonio_aux_locs}
}
