% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{gonio}
\alias{gonio}
\title{Descriptive data title}
\format{
a \code{spec_tbl_df} containing the following fields:
\describe{
\item{hex}{}
\item{date}{}
\item{datenum}{}
\item{gonio_bearing_deg}{}
\item{strength_db}{}
\item{platform}{}
\item{ptt}{}
\item{DeployID}{}
\item{lon}{}
\item{lat}{}
\item{boat_bearing_deg}{}
\item{boat_speed_kph}{}
\item{gps_gonio_time_difference_s}{}
\item{date_time}{}
\item{gap}{}
\item{group}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{BRSPackage}
\link{aux}
\link{processed_gonio_aux_locs}
}
