% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{processed_gonio_aux_locs}
\alias{processed_gonio_aux_locs}
\title{Descriptive data title}
\format{
a \code{grouped_df} containing the following fields:
\describe{
\item{DeployID}{}
\item{Latitude}{}
\item{Longitude}{}
\item{Error.Semi.major.axis}{}
\item{Quality}{}
\item{Source}{}
\item{Error.Semi.minor.axis}{}
\item{Error.Ellipse.orientation}{}
\item{date_time}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{BRSPackage}
\link{aux}
\link{gonio}
}
