% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{series}
\alias{series}
\title{Descriptive data title}
\format{
a \code{spec_tbl_df} containing the following fields:
\describe{
\item{X1}{}
\item{DeployID}{}
\item{Ptt}{}
\item{DepthSensor}{}
\item{Source}{}
\item{Instr}{}
\item{Day}{}
\item{Time}{}
\item{LocationQuality}{}
\item{Latitude}{}
\item{Longitude}{}
\item{Depth}{}
\item{DRange}{}
\item{Temperature}{}
\item{TRange}{}
\item{Activity}{}
\item{ARange}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{BRSPackage}
}
