% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{series_range}
\alias{series_range}
\title{Depth bins of whales}
\format{
a \code{data.frame} containing the following fields:
\describe{
\item{DeployID}{The ID identifying an individual whale.}
\item{Ptt}{The number identifying an individual whale.}
\item{DepthSensor}{The depth of each observation.}
\item{Source}{Whether an observation is from locations, auxiliary, or goniometer data.}
\item{Instr}{Instrument used to identify an individual whale.}
\item{Count}{}
\item{Start}{The date of the first observation from an individual whale.}
\item{End}{The date of the last observation from an individual whale.}
\item{LocationQuality}{N/A}
\item{Latitude}{N/A}
\item{Longitude}{N/A}
\item{MinDepth}{The minimum depth of each observation from an individual whale.}
\item{MinDepthAccuracy}{The accuracy of minimum depth of each each observation.}
\item{MaxDepth}{The maximum depth of each observation from an individual whale.}
\item{MaxDepthAccuracy}{The accuracy of maximum depth of each each observation.}
}
}
\source{
The data comes from the 2019 BRS Study.
}
\description{
Detailed description of the data
}
\seealso{

}
