% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRSPackage.R
\docType{data}
\name{series_xts}
\alias{series_xts}
\title{Descriptive data title}
\format{
a \code{data.frame} containing the following fields:
\describe{
\item{DeployID}{}
\item{start_date}{}
\item{end_date}{}
\item{num_fixes}{}
\item{date_range}{}
\item{average_locs}{}
\item{series_start_date}{}
\item{series_end_date}{}
\item{series_date_range}{}
\item{num_before_cee}{}
\item{num_after_cee}{}
}
}
\source{
The data comes from the 2019 BRS Study.
}
\description{
Detailed description of the data
}
\seealso{

}
